+++ contracts/terraswap_factory/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "terraswap-factory"
-version = "1.0.0"
+version = "1.1.0"
 authors = ["Terraform Labs, PTE."]
 edition = "2018"
 description = "A Terraswap factory contract - auto pair contract generator and also directory for all pairs"
@@ -37,7 +37,7 @@ cranelift = ["cosmwasm-vm/default-cranelift"]
 singlepass = ["cosmwasm-vm/default-singlepass"]
 
 [dependencies]
-terraswap = { path = "../../packages/terraswap", default-features = false, version = "1.0.0"}
+terraswap = { path = "../../packages/terraswap", default-features = false, version = "1.1.0"}
 cosmwasm-std = { version = "0.10.1", features = ["iterator"] }
 cosmwasm-storage = { version = "0.10.1", features = ["iterator"] }
 schemars = "0.7"
