+++ contracts/terraswap_factory/examples/schema.rs
@@ -4,8 +4,8 @@ use std::fs::create_dir_all;
 use cosmwasm_schema::{export_schema, remove_schemas, schema_for};
 use cosmwasm_std::HandleResponse;
 
-use terraswap::PairInfo;
-use terraswap_factory::msg::{ConfigResponse, HandleMsg, InitMsg, QueryMsg, PairsResponse};
+use terraswap::asset::PairInfo;
+use terraswap::factory::{ConfigResponse, HandleMsg, InitMsg, PairsResponse, QueryMsg};
 
 fn main() {
     let mut out_dir = current_dir().unwrap();
