+++ contracts/terraswap_factory/src/lib.rs
@@ -1,5 +1,4 @@
 pub mod contract;
-pub mod msg;
 pub mod state;
 
 mod querier;
