+++ contracts/terraswap_pair/examples/schema.rs
@@ -4,9 +4,10 @@ use std::fs::create_dir_all;
 use cosmwasm_schema::{export_schema, remove_schemas, schema_for};
 use cosmwasm_std::HandleResponse;
 
-use terraswap::{PairInfo, PairInitMsg};
-use terraswap_pair::msg::{
-    Cw20HookMsg, HandleMsg, PoolResponse, QueryMsg, ReverseSimulationResponse, SimulationResponse,
+use terraswap::asset::PairInfo;
+use terraswap::pair::{
+    Cw20HookMsg, HandleMsg, InitMsg, PoolResponse, QueryMsg, ReverseSimulationResponse,
+    SimulationResponse,
 };
 
 fn main() {
@@ -15,7 +16,7 @@ fn main() {
     create_dir_all(&out_dir).unwrap();
     remove_schemas(&out_dir).unwrap();
 
-    export_schema(&schema_for!(PairInitMsg), &out_dir);
+    export_schema(&schema_for!(InitMsg), &out_dir);
     export_schema(&schema_for!(HandleMsg), &out_dir);
     export_schema(&schema_for!(Cw20HookMsg), &out_dir);
     export_schema(&schema_for!(HandleResponse), &out_dir);
