+++ contracts/terraswap_pair/schema/init_msg.json
@@ -1,45 +1,39 @@
 {
   "$schema": "http://json-schema.org/draft-07/schema#",
-  "title": "FactoryHandleMsg",
-  "anyOf": [
-    {
-      "description": "CreatePair instantiates pair contract",
-      "type": "object",
-      "required": [
-        "create_pair"
-      ],
-      "properties": {
-        "create_pair": {
-          "type": "object",
-          "required": [
-            "asset_infos"
-          ],
-          "properties": {
-            "asset_infos": {
-              "description": "Asset infos",
-              "type": "array",
-              "items": {
-                "$ref": "#/definitions/AssetInfo"
-              },
-              "maxItems": 2,
-              "minItems": 2
-            },
-            "init_hook": {
-              "description": "Init hook for after works",
-              "anyOf": [
-                {
-                  "$ref": "#/definitions/InitHook"
-                },
-                {
-                  "type": "null"
-                }
-              ]
-            }
-          }
-        }
-      }
-    }
+  "title": "InitMsg",
+  "type": "object",
+  "required": [
+    "asset_infos",
+    "token_code_id"
   ],
+  "properties": {
+    "asset_infos": {
+      "description": "Asset infos",
+      "type": "array",
+      "items": {
+        "$ref": "#/definitions/AssetInfo"
+      },
+      "maxItems": 2,
+      "minItems": 2
+    },
+    "init_hook": {
+      "description": "Hook for post initalization",
+      "anyOf": [
+        {
+          "$ref": "#/definitions/InitHook"
+        },
+        {
+          "type": "null"
+        }
+      ]
+    },
+    "token_code_id": {
+      "description": "Token contract code id for initialization",
+      "type": "integer",
+      "format": "uint64",
+      "minimum": 0.0
+    }
+  },
   "definitions": {
     "AssetInfo": {
       "anyOf": [
