--- contracts/terraswap_pair/src/msg.rs
@@ -1,74 +1,0 @@
-use schemars::JsonSchema;
-use serde::{Deserialize, Serialize};
-
-use cosmwasm_std::{Decimal, HumanAddr, Uint128};
-use cw20::Cw20ReceiveMsg;
-use terraswap::Asset;
-
-#[derive(Serialize, Deserialize, Clone, Debug, PartialEq, JsonSchema)]
-#[serde(rename_all = "snake_case")]
-pub enum HandleMsg {
-    Receive(Cw20ReceiveMsg),
-    /// Post initize step to allow user to set controlled contract address after creating it
-    PostInitialize {},
-    /// ProvideLiquidity a user provides pool liquidity
-    ProvideLiquidity {
-        assets: [Asset; 2],
-        slippage_tolerance: Option<Decimal>,
-    },
-    /// Swap an offer asset to the other
-    Swap {
-        offer_asset: Asset,
-        belief_price: Option<Decimal>,
-        max_spread: Option<Decimal>,
-        to: Option<HumanAddr>,
-    },
-}
-
-#[derive(Serialize, Deserialize, Clone, Debug, PartialEq, JsonSchema)]
-#[serde(rename_all = "snake_case")]
-pub enum Cw20HookMsg {
-    /// Sell a given amount of asset
-    Swap {
-        belief_price: Option<Decimal>,
-        max_spread: Option<Decimal>,
-        to: Option<HumanAddr>,
-    },
-    WithdrawLiquidity {},
-}
-
-#[derive(Serialize, Deserialize, Clone, Debug, PartialEq, JsonSchema)]
-#[serde(rename_all = "snake_case")]
-pub enum QueryMsg {
-    Pair {},
-    Pool {},
-    Simulation { offer_asset: Asset },
-    ReverseSimulation { ask_asset: Asset },
-}
-
-// We define a custom struct for each query response
-#[derive(Serialize, Deserialize, Clone, Debug, PartialEq, JsonSchema)]
-pub struct PoolResponse {
-    pub assets: [Asset; 2],
-    pub total_share: Uint128,
-}
-
-/// SimulationResponse returns swap simulation response
-#[derive(Serialize, Deserialize, Clone, Debug, PartialEq, JsonSchema)]
-pub struct SimulationResponse {
-    pub return_amount: Uint128,
-    pub spread_amount: Uint128,
-    pub commission_amount: Uint128,
-}
-
-/// ReverseSimulationResponse returns reverse swap simulation response
-#[derive(Serialize, Deserialize, Clone, Debug, PartialEq, JsonSchema)]
-pub struct ReverseSimulationResponse {
-    pub offer_amount: Uint128,
-    pub spread_amount: Uint128,
-    pub commission_amount: Uint128,
-}
-
-/// We currently take no arguments for migrations
-#[derive(Serialize, Deserialize, Clone, Debug, PartialEq, JsonSchema)]
-pub struct MigrateMsg {}
