+++ contracts/terraswap_token/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "terraswap-token"
-version = "1.0.0"
+version = "1.1.0"
 authors = ["Terraform Labs, PTE."]
 edition = "2018"
 description = "Expanded implementation of a CosmWasm-20 compliant token for post intialization"
@@ -18,7 +18,7 @@ backtraces = ["cosmwasm-std/backtraces"]
 library = []
 
 [dependencies]
-terraswap = { path = "../../packages/terraswap", default-features = false, version = "1.0.0"}
+terraswap = { path = "../../packages/terraswap", default-features = false, version = "1.1.0"}
 cw2 = "0.2"
 cw20 = "0.2"
 cw20-base = { version = "0.2", features = ["library"] }
