+++ contracts/terraswap_token/examples/schema.rs
@@ -8,7 +8,7 @@ use cw20::{
     TokenInfoResponse,
 };
 use cw20_base::msg::{HandleMsg, MigrateMsg, QueryMsg};
-use terraswap::{TokenCw20HookMsg, TokenInitMsg};
+use terraswap::token::InitMsg;
 
 fn main() {
     let mut out_dir = current_dir().unwrap();
@@ -16,8 +16,7 @@ fn main() {
     create_dir_all(&out_dir).unwrap();
     remove_schemas(&out_dir).unwrap();
 
-    export_schema(&schema_for!(TokenInitMsg), &out_dir);
-    export_schema(&schema_for!(TokenCw20HookMsg), &out_dir);
+    export_schema(&schema_for!(InitMsg), &out_dir);
     export_schema(&schema_for!(HandleMsg), &out_dir);
     export_schema(&schema_for!(QueryMsg), &out_dir);
     export_schema(&schema_for!(MigrateMsg), &out_dir);
