+++ contracts/terraswap_token/schema/init_msg.json
@@ -1,6 +1,6 @@
 {
   "$schema": "http://json-schema.org/draft-07/schema#",
-  "title": "TokenInitMsg",
+  "title": "InitMsg",
   "description": "TokenContract InitMsg",
   "type": "object",
   "required": [
