--- contracts/terraswap_token/schema/token_cw20_hook_msg.json
@@ -1,17 +1,0 @@
-{
-  "$schema": "http://json-schema.org/draft-07/schema#",
-  "title": "TokenCw20HookMsg",
-  "anyOf": [
-    {
-      "type": "object",
-      "required": [
-        "migrate"
-      ],
-      "properties": {
-        "migrate": {
-          "type": "object"
-        }
-      }
-    }
-  ]
-}
