+++ packages/terraswap/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "terraswap"
-version = "1.0.1"
+version = "1.1.0"
 authors = ["Terraform Labs, PTE."]
 edition = "2018"
 description = "Common terraswap types"
