--- packages/terraswap/schema/factory_query_msg.json
@@ -1,77 +1,0 @@
-{
-  "$schema": "http://json-schema.org/draft-07/schema#",
-  "title": "FactoryQueryMsg",
-  "anyOf": [
-    {
-      "type": "object",
-      "required": [
-        "pair"
-      ],
-      "properties": {
-        "pair": {
-          "type": "object",
-          "required": [
-            "asset_infos"
-          ],
-          "properties": {
-            "asset_infos": {
-              "type": "array",
-              "items": {
-                "$ref": "#/definitions/AssetInfo"
-              },
-              "maxItems": 2,
-              "minItems": 2
-            }
-          }
-        }
-      }
-    }
-  ],
-  "definitions": {
-    "AssetInfo": {
-      "anyOf": [
-        {
-          "type": "object",
-          "required": [
-            "token"
-          ],
-          "properties": {
-            "token": {
-              "type": "object",
-              "required": [
-                "contract_addr"
-              ],
-              "properties": {
-                "contract_addr": {
-                  "$ref": "#/definitions/HumanAddr"
-                }
-              }
-            }
-          }
-        },
-        {
-          "type": "object",
-          "required": [
-            "native_token"
-          ],
-          "properties": {
-            "native_token": {
-              "type": "object",
-              "required": [
-                "denom"
-              ],
-              "properties": {
-                "denom": {
-                  "type": "string"
-                }
-              }
-            }
-          }
-        }
-      ]
-    },
-    "HumanAddr": {
-      "type": "string"
-    }
-  }
-}
