--- packages/terraswap/schema/pair_query_msg.json
@@ -1,109 +1,0 @@
-{
-  "$schema": "http://json-schema.org/draft-07/schema#",
-  "title": "PairQueryMsg",
-  "anyOf": [
-    {
-      "type": "object",
-      "required": [
-        "simulation"
-      ],
-      "properties": {
-        "simulation": {
-          "type": "object",
-          "required": [
-            "offer_asset"
-          ],
-          "properties": {
-            "offer_asset": {
-              "$ref": "#/definitions/Asset"
-            }
-          }
-        }
-      }
-    },
-    {
-      "type": "object",
-      "required": [
-        "reverse_simulation"
-      ],
-      "properties": {
-        "reverse_simulation": {
-          "type": "object",
-          "required": [
-            "ask_asset"
-          ],
-          "properties": {
-            "ask_asset": {
-              "$ref": "#/definitions/Asset"
-            }
-          }
-        }
-      }
-    }
-  ],
-  "definitions": {
-    "Asset": {
-      "type": "object",
-      "required": [
-        "amount",
-        "info"
-      ],
-      "properties": {
-        "amount": {
-          "$ref": "#/definitions/Uint128"
-        },
-        "info": {
-          "$ref": "#/definitions/AssetInfo"
-        }
-      }
-    },
-    "AssetInfo": {
-      "anyOf": [
-        {
-          "type": "object",
-          "required": [
-            "token"
-          ],
-          "properties": {
-            "token": {
-              "type": "object",
-              "required": [
-                "contract_addr"
-              ],
-              "properties": {
-                "contract_addr": {
-                  "$ref": "#/definitions/HumanAddr"
-                }
-              }
-            }
-          }
-        },
-        {
-          "type": "object",
-          "required": [
-            "native_token"
-          ],
-          "properties": {
-            "native_token": {
-              "type": "object",
-              "required": [
-                "denom"
-              ],
-              "properties": {
-                "denom": {
-                  "type": "string"
-                }
-              }
-            }
-          }
-        }
-      ]
-    },
-    "HumanAddr": {
-      "type": "string"
-    },
-    "Uint128": {
-      "type": "string"
-    }
-  }
-}
