+++ packages/terraswap/src/hook.rs
@@ -7,9 +7,3 @@ pub struct InitHook {
     pub msg: Binary,
     pub contract_addr: HumanAddr,
 }
-
-#[derive(Serialize, Deserialize, Clone, Debug, PartialEq, JsonSchema)]
-#[serde(rename_all = "snake_case")]
-pub enum TokenCw20HookMsg {
-    Migrate {},
-}
