+++ packages/terraswap/src/lib.rs
@@ -1,19 +1,9 @@
-mod asset;
-mod hook;
-mod init;
-mod msg;
-mod querier;
-
-pub use crate::asset::{Asset, AssetInfo, AssetInfoRaw, AssetRaw, PairInfo, PairInfoRaw};
-pub use crate::hook::{InitHook, TokenCw20HookMsg};
-pub use crate::init::{PairInitMsg, TokenInitMsg};
-pub use crate::msg::{
-    FactoryHandleMsg, FactoryQueryMsg, PairCw20HookMsg, PairHandleMsg, PairQueryMsg,
-};
-pub use crate::querier::{
-    query_all_balances, query_balance, query_pair_info, query_supply, query_token_balance,
-    reverse_simulate, simulate,
-};
+pub mod asset;
+pub mod factory;
+pub mod hook;
+pub mod pair;
+pub mod querier;
+pub mod token;
 
 #[cfg(test)]
 mod mock_querier;
